Update my fork

+++ fhem/FHEM/70_XBMC.pm
@@ -29,7 +29,8 @@ sub XBMC_Initialize($$)
   $hash->{ReadFn}   = "XBMC_Read";  
   $hash->{ReadyFn}  = "XBMC_Ready";
   $hash->{UndefFn}  = "XBMC_Undefine";
-  $hash->{AttrList} = "fork:enable,disable compatibilityMode:xbmc,plex offMode:quit,hibernate,shutdown,standby updateInterval " . $readingFnAttributes;
+  $hash->{AttrFn}   = "XBMC_Attr";
+  $hash->{AttrList} = "fork:enable,disable compatibilityMode:xbmc,plex offMode:quit,hibernate,shutdown,standby updateInterval disable:1,0 " . $readingFnAttributes;
   
   $data{RC_makenotify}{XBMC} = "XBMC_RCmakenotify";
   $data{RC_layout}{XBMC_RClayout}  = "XBMC_RClayout";
@@ -45,6 +46,7 @@ sub XBMC_Define($$)
   }
   my ($name, $type, $addr, $protocol, $username, $password) = @args;
   $hash->{Protocol} = $protocol;
+  $hash->{NextID} = 1;
   $addr =~ /^(.*?)(:([0-9]+))?$/;  
   $hash->{Host} = $1;
   if(defined($3)) {
@@ -75,6 +77,36 @@ sub XBMC_Define($$)
   return undef;
 }
 
+sub XBMC_Attr($$$$)
+{
+  my ($cmd, $name, $attr, $value) = @_;
+  my $hash = $defs{$name};
+  
+  if($attr eq "disable") {
+    if($cmd eq "set" && ($value || !defined($value))) {
+      XBMC_Disconnect($hash);
+      $hash->{STATE} = "Disabled";
+    } else {
+      if (AttrVal($hash->{NAME}, 'disable', 0)) {
+        $hash->{STATE} = "Initialized";
+        
+        my $dev = $hash->{DeviceName};
+        $readyfnlist{"$name.$dev"} = $hash;
+      }
+    }
+  }
+
+  return undef;
+}
+
+sub XBMC_CreateId($) 
+{
+  my ($hash) = @_;
+  my $res = $hash->{NextID};
+  $hash->{NextID} = ($res >= 1000000) ? 1 : $res + 1;
+  return $res;
+}
+
 # Force a connection attempt to XBMC as soon as possible 
 # (e.g. you know you just started it and want to connect immediately without waiting up to 60 s)
 sub XBMC_Connect($)
@@ -99,13 +131,29 @@ sub XBMC_Connect($)
   } else {
     $hash->{NEXT_OPEN} = 0; # force NEXT_OPEN used in DevIO
   }
-    
+
   return undef;
+}
+
+# kills child process trying to connect (if existing)
+sub XBMC_KillConnectionChild($)
+{
+  my ($hash) = @_;
+
+  return if !$hash->{CHILDPID};
+    
+  kill 'KILL', $hash->{CHILDPID};
+  undef $hash->{CHILDPID};
 }
 
 sub XBMC_Ready($)
 {
   my ($hash) = @_;
+  
+  if (AttrVal($hash->{NAME}, 'disable', 0)) {
+    return;
+  }
+  
   if($hash->{Protocol} eq 'tcp') {
     if(AttrVal($hash->{NAME},'fork','disable') eq 'enable') {
       if($hash->{CHILDPID} && !(kill 0, $hash->{CHILDPID})) {
@@ -153,10 +201,19 @@ sub XBMC_Undefine($$)
   
   RemoveInternalTimer($hash);
   
+  XBMC_Disconnect($hash);
+  
+  return undef;
+}
+
+sub XBMC_Disconnect($)
+{
+  my ($hash) = @_;
   if($hash->{Protocol} eq 'tcp') {
     DevIo_CloseDev($hash); 
   }
-  return undef;
+  
+  XBMC_KillConnectionChild($hash);
 }
 
 sub XBMC_Init($) 
@@ -293,7 +350,7 @@ sub XBMC_PlayerGetItem($$)
     "method" => "Player.GetItem",
     "params" => { 
       "properties" => ["artist", "album", "thumbnail", "file", "title",
-                        "track", "year", "streamdetails"]
+                        "track", "year", "streamdetails", "tvshowid"]
     }
   };
   if($playerid >= 0) {    
@@ -349,7 +406,9 @@ sub XBMC_ProcessRead($$)
     }
     #otherwise it is a answer of a request
     else {
-      XBMC_ProcessResponse($hash,$obj);
+        if (XBMC_ProcessResponse($hash,$obj) == -1) {
+            Log3($name, 2, "XBMC_ProcessRead: Faulty message: $msg");
+        }
     }
     ($msg,$tail) = XBMC_ParseMsg($hash, $tail);
   }
@@ -387,7 +446,7 @@ sub XBMC_ResetMediaReadings($)
   # delete streamdetails readings
   # NOTE: we actually delete the readings (unlike the other readings)
   #       because they are stream count dependent
-  fhem("deletereading $hash->{NAME} sd_.*");
+  fhem("deletereading $hash->{NAME} sd_.*", 1);
 }
 
 sub XBMC_ResetPlayerReadings($)
@@ -409,7 +468,7 @@ sub XBMC_PlayerOnPlay($$)
 {
   my ($hash,$obj) = @_;
   my $name = $hash->{NAME};
-  my $id = XBMC_CreateId();
+  my $id = XBMC_CreateId($hash);
   my $type = $obj->{params}->{data}->{item}->{type};
   if(AttrVal($hash->{NAME},'compatibilityMode','xbmc') eq 'plex' || !defined($obj->{params}->{data}->{item}->{id}) || $type eq "picture" || $type eq "unknown") {
     # we either got unknown or picture OR an item not in the library (id not existing)
@@ -569,6 +628,7 @@ sub XBMC_ProcessNotification($$)
 sub XBMC_ProcessResponse($$) 
 {
   my ($hash,$obj) = @_;
+  my $name = $hash->{NAME};
   my $id = $obj->{id};
   #check if the id of the answer matches the id of a pending event
   if(defined($hash->{PendingEvents}{$id})) {
@@ -593,15 +653,25 @@ sub XBMC_ProcessResponse($$)
     } 
     $hash->{PendingEvents}{$id} = undef;
   }
-  elsif(defined($hash->{PendingPlayerCMDs}{$id})) {
+  elsif(exists($hash->{PendingPlayerCMDs}{$id})) {
     my $cmd = $hash->{PendingPlayerCMDs}{$id};
     my $players = $obj->{result};
+    if (ref($players) ne "ARRAY") {
+        my $keys = "";
+        while ((my $k, my $v) = each $hash->{PendingPlayerCMDs} ) {
+          $keys .= ",$k";
+        }
+        delete $hash->{PendingPlayerCMDs}{$id};
+        Log3($name, 2, "XBMC_ProcessResponse: Not received a player array! Pending command cancelled!");
+        Log3($name, 2, "XBMC_ProcessResponse: Keys in PendingPlayerCMDs: $keys");
+        return -1;
+    }
     foreach my $player (@$players) {
-      $cmd->{id} = XBMC_CreateId();
+      $cmd->{id} = XBMC_CreateId($hash);
       $cmd->{params}->{playerid} = $player->{playerid};
       XBMC_Call($hash,$cmd,1);
     }
-    $hash->{PendingPlayerCMDs}{$id} = undef;
+    delete $hash->{PendingPlayerCMDs}{$id};
   }  
   else {
     my $result = $obj->{result};
@@ -623,7 +693,7 @@ sub XBMC_ProcessResponse($$)
       readingsEndUpdate($hash, 1);
     }
   }
-  return undef;
+  return 0;
 }
 
 sub XBMC_Is3DFile($$) {
@@ -787,6 +857,9 @@ sub XBMC_Set($@)
   }
   elsif($cmd eq 'openepisodeid') {
     return XBMC_Set_Open($hash, 'episode', @args);
+  }
+  elsif($cmd eq 'openchannelid') {
+    return XBMC_Set_Open($hash, 'channel', @args);
   }
   elsif($cmd eq 'addon') {
     return XBMC_Set_Addon($hash, @args);
@@ -992,6 +1065,12 @@ sub XBMC_Set_Open($@)
         'resume' => JSON::true
        }
     };
+  } elsif($opt eq 'channel') {
+    $params = { 
+      'item' => {
+        'channelid' => $path +0
+      },
+    };
   }
   my $obj = {
     'method' => 'Player.Open',
@@ -1172,7 +1251,7 @@ sub XBMC_PlayerCommand($$$)
   }
   
   #we need to find out the correct player first
-  my $id = XBMC_CreateId();
+  my $id = XBMC_CreateId($hash);
   $hash->{PendingPlayerCMDs}->{$id} = $obj;
   my $req = {
     'method'  => 'Player.GetActivePlayers',
@@ -1215,7 +1294,7 @@ sub XBMC_Call($$$)
   my $name = $hash->{NAME};
   #add an ID otherwise XBMC will not respond
   if($id &&!defined($obj->{id})) {
-    $obj->{id} = XBMC_CreateId();
+    $obj->{id} = XBMC_CreateId($hash);
   }
   $obj->{jsonrpc} = "2.0"; #JSON RPC version has to be passed
   my $json = JSON->new->utf8(0)->encode($obj);
@@ -1239,11 +1318,6 @@ sub XBMC_Call_raw($$$)
   else {
     return XBMC_TCP_Call($hash,$obj);
   }
-}
-
-sub XBMC_CreateId() 
-{
-  return int(rand(1000000));
 }
 
 sub XBMC_RCmakenotify($$) {
@@ -1466,8 +1540,9 @@ sub XBMC_HTTP_Request($$@)
     <li><b>repeat &lt;one|all|off&gt; [&lt;audio|video|picture&gt;]</b> -  Sets the repeat mode.</li>
     <li><b>open &lt;URI&gt;</b> -  Plays the resource located at the URI (can be a url or a file)</li>
     <li><b>opendir &lt;path&gt;</b> -  Plays the content of the directory</li>
-    <li><b>openmovieid &lt;path&gt;</b> -  Plays the movie of the id</li>
-    <li><b>openepisodeid &lt;path&gt;</b> -  Plays the episode of the id</li>
+    <li><b>openmovieid &lt;path&gt;</b> -  Plays a movie by id</li>
+    <li><b>openepisodeid &lt;path&gt;</b> -  Plays an episode by id</li>
+    <li><b>openchannelid &lt;path&gt;</b> -  Switches to channel by id</li>
     <li><b>addon &lt;addonid&gt; &lt;parametername&gt; &lt;parametervalue&gt;</b> -  Executes addon with one Parameter, for example set xbmc addon script.json-cec command activate</li>
   </ul>
   <br>Input related commands:<br>
@@ -1588,8 +1663,10 @@ sub XBMC_HTTP_Request($$@)
       If XBMC does not run all the time it used to be the case that FHEM blocks because it cannot reach XBMC (only happened 
     if TCP was used). If you encounter problems like FHEM not responding for a few seconds then you should set <code>attr &lt;XBMC_device&gt; fork enable</code>
     which will move the search for XBMC into a separate process.</li>
-    <li>updateInterval<br>
+  <li>updateInterval<br>
       The interval which is used to check if Kodi is still alive (by sending a JSON ping) and also it is used to update current player item.</li>
+  <li>disable<br>
+      Disables the device. All connections will be closed immediately.</li>
   </ul>
 </ul>
